logclient: move ws list to separate file
