+++ pkg/util/logclient/consts.go
@@ -14,79 +14,6 @@
 
 package logclient
 
-// 这些状态不做 websocket 通知
-var BlackList = []string{
-	ACT_CLOUD_FULLSYNC,
-	ACT_CREATE,
-	ACT_DELETE,
-	ACT_PENDING_DELETE,
-	ACT_PRIVATE,
-	ACT_PUBLIC,
-	ACT_UPDATE,
-	ACT_VM_SYNC_STATUS,
-	ACT_VM_SYNC_CONF,
-}
-
-// 这些状态需要做 websocket 通知
-var WhiteList = []string{
-	ACT_ADDTAG,
-	ACT_ALLOCATE,
-	ACT_DELOCATE,
-	ACT_BM_CONVERT_HYPER,
-	ACT_BM_MAINTENANCE,
-	ACT_BM_UNCONVERT_HYPER,
-	ACT_BM_UNMAINTENANCE,
-	ACT_CANCEL_DELETE,
-	ACT_CHANGE_OWNER,
-	ACT_SYNC_CLOUD_OWNER,
-	ACT_CLOUD_SYNC,
-	ACT_DISABLE,
-	ACT_ENABLE,
-	ACT_GUEST_ATTACH_ISOLATED_DEVICE,
-	ACT_GUEST_DETACH_ISOLATED_DEVICE,
-	ACT_MERGE,
-	ACT_OFFLINE,
-	ACT_ONLINE,
-	ACT_RELEASE_IP,
-	ACT_RESERVE_IP,
-	ACT_RESIZE,
-	ACT_RMTAG,
-	ACT_SPLIT,
-	ACT_UNCACHED_IMAGE,
-	ACT_VM_ATTACH_DISK,
-	ACT_VM_BIND_KEYPAIR,
-	ACT_VM_CHANGE_FLAVOR,
-	ACT_VM_DEPLOY,
-	ACT_VM_DETACH_DISK,
-	ACT_VM_PURGE,
-	ACT_VM_REBUILD,
-	ACT_VM_RESET_PSWD,
-	ACT_VM_CHANGE_BANDWIDTH,
-	ACT_VM_START,
-	ACT_VM_STOP,
-	ACT_VM_RESTART,
-	ACT_VM_UNBIND_KEYPAIR,
-	ACT_VM_ASSIGNSECGROUP,
-	ACT_VM_REVOKESECGROUP,
-	ACT_VM_SETSECGROUP,
-	ACT_RESET_DISK,
-	ACT_SYNC_STATUS,
-	ACT_SYNC_CONF,
-	ACT_CREATE_BACKUP,
-	ACT_SWITCH_TO_BACKUP,
-	ACT_RENEW,
-	ACT_MIGRATE,
-	ACT_IMAGE_SAVE,
-	ACT_RECYCLE_PREPAID,
-	ACT_UNDO_RECYCLE_PREPAID,
-	ACT_FETCH,
-	ACT_VM_CHANGE_NIC,
-	ACT_HOST_IMPORT_LIBVIRT_SERVERS,
-	ACT_GUEST_CREATE_FROM_IMPORT,
-	ACT_DISK_CREATE_SNAPSHOT,
-	ACT_IMAGE_PROBE,
-}
-
 const (
 	ACT_ADDTAG                       = "添加标签"
 	ACT_ALLOCATE                     = "分配"
